PdfOppretterService struktur
